feat: add error handling for deleteProduct method. Check if product exist nad that the owner (or admin) of the product can only delete the product

+++ src/services/repositories/product.repository.js
@@ -1,7 +1,8 @@
+import { AppError } from "../../helpers/AppError.js";
 import { generateProduct } from "../../utils/faker.js";
+import { EErrors } from "../errors/enums.js";
 
 export default class ProductRepository {
-
   constructor(dao) {
     this.dao = dao;
   }
@@ -16,6 +17,7 @@ export default class ProductRepository {
       status: body.status,
       stock: body.stock,
       thumbnails: body.thumbnails,
+      owner: body.owner
     });
   };
 
@@ -23,11 +25,28 @@ export default class ProductRepository {
     return await this.dao.findByIdAndUpdate(id, upFields);
   };
 
-  updateProductStock = async (id,newStock) => {
-    return await this.dao.updateStock(id,newStock)
-  }
+  updateProductStock = async (id, newStock) => {
+    return await this.dao.updateStock(id, newStock);
+  };
 
-  deleteProduct = async (id) => {
+  deleteProduct = async (id, user) => {
+    const foundProduct = await this.dao.findById(id);
+    if (!foundProduct) {
+      throw new AppError({
+        name: "Delete product error.",
+        message: "Error while trying to delete product.",
+        code: EErrors.NOT_FOUND,
+        cause: `The product you are trying to delete does not exist!`,
+      });
+    }
+    if (user.role !== "admin" && foundProduct.owner !== user.email) {
+      throw new AppError({
+        name: "Delete product error.",
+        message: "Error while trying to delete product.",
+        code: EErrors.UNAUTHORIZED,
+        cause: `You are not authorized to delete this product.`,
+      });
+    }
     return await this.dao.deleteOneById(id);
   };
 
@@ -46,17 +65,17 @@ export default class ProductRepository {
       {
         limit,
         page,
-        sort
+        sort,
       }
     );
   };
 
   getMockingProducts = async (amount = 100) => {
-    const mockingproducts = []
+    const mockingproducts = [];
     for (let i = 0; i < amount; i++) {
-      const generatedProduct = generateProduct()
-      mockingproducts.push(generatedProduct)
+      const generatedProduct = generateProduct();
+      mockingproducts.push(generatedProduct);
     }
-    return mockingproducts
-  }
+    return mockingproducts;
+  };
 }
